+++ src/WhatBins/BinCollectionsFinder.cs
@@ -1,0 +1,12 @@
+﻿namespace WhatBins
+{
+    using System;
+
+    public class BinCollectionsFinder
+    {
+        public LookupResult Lookup(PostCode postCode)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
