+++ src/WhatBins/Collection.cs
@@ -1,0 +1,6 @@
+﻿namespace WhatBins
+{
+    public class Collection
+    {
+    }
+}
