+++ src/WhatBins/CollectionState.cs
@@ -1,0 +1,9 @@
+﻿namespace WhatBins
+{
+    public enum CollectionState
+    {
+        Unsupported,
+        NoCollection,
+        Collection
+    }
+}
