+++ src/WhatBins/PostCode.cs
@@ -1,0 +1,13 @@
+﻿namespace WhatBins
+{
+    public struct PostCode
+    {
+        private string value;
+
+        public PostCode(string value)
+            : this()
+        {
+            this.value = value;
+        }
+    }
+}
