+++ tests/WhatBins.IntegrationTests/WhatBins.IntegrationTests.csproj
@@ -24,4 +24,8 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\WhatBins\WhatBins.csproj" />
+  </ItemGroup>
+
 </Project>
